--- cashier/chefTest.php
@@ -1,15 +1,0 @@
-<html>
-    <head>
-        
-    </head>
-    <body>
-        <h1>This is the Chef Homepage</h1>
-
-    </body>
-</html><?php
-
-/* 
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Scripting/EmptyPHP.php to edit this template
- */
-
